Version 0.4.6.5 - xsendfile extension not used when sending error code
The nginx's xsendfile extension allows to deliver static content
directly instead of passing it through the proxied application.
Unfortunately if we want to return a 404 code then nginx will deliver
the image with code 200 (or 304) and ignore our 404 return code.
So now we are not using nginx's extension in case of 404s.